+++ Data/Aeson/AutoType/Extract.hs
@@ -5,8 +5,10 @@ module Data.Aeson.AutoType.Extract(valueSize, valueTypeSize,
                                    extractType, unifyTypes,
                                    typeCheck) where
 
+import           Control.Arrow ((&&&))
 import           Control.Exception               (assert)
 import           Data.Aeson.AutoType.Type
+import qualified Data.Graph          as Graph
 import qualified Data.HashMap.Strict      as Map
 import           Data.HashMap.Strict             (HashMap)
 import qualified Data.Set                 as Set
@@ -152,3 +154,4 @@ simplifyUnion (TUnion s)                   = TUnion $ Set.unions $ map elements 
     elements (TUnion elems) = elems
     elements sing           = Set.singleton sing
 simplifyUnion unexpected                   = error ("simplifyUnion: unexpected argument " ++ show unexpected)
+